Protect `DisciplineDetailView`
The class-based `DisciplineDetailView` in the lifestudies module
has been protected using the GroupRequiredMixin Django feature.

This commit makes it so only Users in the training_assitant group
can access this view.
A "Forbidden" message will appear otherwise if a User who is not
a in the group tries to access the view.

Fixes: #1323